fix(gnovm): drop an unnecessary conversion in banker_test
btReadonly is declared uint8, so the whole const block is uint8 and
uint8(btOriginSend) converts a value to its own type. golangci-lint's
unconvert flags it, and CI on PR 6095 fails on it.

Mine, from the test added earlier in this branch. I ran the test suites
every time and never ran the linter.

+++ gnovm/stdlibs/chain/banker/banker_test.go
@@ -49,7 +49,7 @@ func TestOriginSendIsSpendableOnlyByThePaidRealm(t *testing.T) {
 				panicked = true
 			}
 		}()
-		X_bankerSendCoins(m, uint8(btOriginSend), string(from), string(to),
+		X_bankerSendCoins(m, btOriginSend, string(from), string(to),
 			[]string{"ugnot"}, []int64{10})
 		return false
 	}
